pyproject.toml
[tool.poetry]
name = "giant-newsletter"
version = "0.3"
description = "A small reusable package that adds a Newsletter app to a project"
authors = ["Will-Hoey <will.hoey@giantmade.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/giantmade/giant-newsletter"
repository = "https://github.com/giantmade/giant-newsletter"
keywords = ["newsletter", "app"]
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

include = [
    "LICENSE",
]
packages = [
    { include = "newsletter" }
]

[tool.poetry.dependencies]
python = "^3.6"
giant-mixins = "^0.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
django = "2.2"
django-cms = "^3.7.2"
black = "^19.10b0"
pytest-cov = "^2.10.1"
pytest-django = "^3.10.0"
six = "^1.15.0"

[[tool.poetry.source]]
name = "TestPyPi"
url = "https://test.pypi.org/simple/"
secondary = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
